Added fix and deleted unneeded items
