--- .vercel/output/builds.json
@@ -1,18 +1,0 @@
-{
-  "//": "This file was generated by the `vercel build` command. It is not part of the Build Output API.",
-  "target": "preview",
-  "argv": [
-    "/opt/homebrew/Cellar/node/23.7.0/bin/node",
-    "/opt/homebrew/bin/vercel",
-    "build"
-  ],
-  "builds": [
-    {
-      "require": "@vercel/node",
-      "requirePath": "/opt/homebrew/lib/node_modules/vercel/node_modules/@vercel/node/dist/index",
-      "apiVersion": 3,
-      "src": "src/index.js",
-      "use": "@vercel/node"
-    }
-  ]
-}
