--- .vercel/output/config.json
@@ -1,10 +1,0 @@
-{
-  "version": 3,
-  "routes": [
-    {
-      "src": "^/(.*)$",
-      "dest": "src/index.js"
-    }
-  ],
-  "crons": []
-}
