--- .vercel/output/diagnostics/cli_traces.json
@@ -1,1 +1,0 @@
-[{"name":"vc.builder","duration":3490599,"timestamp":1894167958840,"id":"73e41f1b-a2f3-47ee-8792-40f2de6cd1f1","parentId":"6168500b-64bc-4d45-8bc8-674fbf7cfca5","tags":{"name":"@vercel/node"},"startTime":1743002963890},{"name":"vc.builder.diagnostics","duration":13,"timestamp":1894171449509,"id":"3c67bdc8-3b06-41e0-b52d-365dc8eed44b","parentId":"73e41f1b-a2f3-47ee-8792-40f2de6cd1f1","tags":{},"startTime":1743002967381},{"name":"vc.doBuild","duration":3929239,"timestamp":1894167685452,"id":"6168500b-64bc-4d45-8bc8-674fbf7cfca5","parentId":"ce6a8d5a-7e65-4b95-8cc4-ac10966fe725","tags":{},"startTime":1743002963617},{"name":"vc","duration":25918677,"timestamp":1894145696023,"id":"ce6a8d5a-7e65-4b95-8cc4-ac10966fe725","tags":{},"startTime":1743002941627}]
