--- .vercel/output/functions/src/index.js.func/.env
@@ -1,8 +1,0 @@
-PORT=6002
-DATABASE_URL=mongodb+srv://jaydenowenswork:testingNBA@cluster0.ufqo1.mongodb.net/
-STRIPE_SECRET_KEY=sk_test_51PIqNl2MgcnDTdrwWh7wv1O7CsMG61p3kbDXrFsUwJJtwXB83WmptGzOiz6RbK1YUUs2Gkv7Y9vCDGZ2JEkInGIf00SwEdbeI3
-STRIPE_WEBHOOK_SECRET=whsec_c2bd43a506899d715d5034fd9efe05b926e3b00bfd8f5ad78c85aa124b0a3e7d
-CHARGEBEE_API_KEY=live_oeUe1fGVscunCF53KaDDUIo5egJbAcuATq
-CHARGEBEE_STRIPE_GATEWAY_ID=gw_16CcWVUVuUSzbpp9
-CHARGEBEE_SITE=thefantasyhacker
-JWT_SECRET=MySecretKey
