--- .vercel/output/functions/src/index.js.func/.vc-config.json
@@ -1,9 +1,0 @@
-{
-  "handler": "src/index.js",
-  "runtime": "nodejs22.x",
-  "environment": {},
-  "launcherType": "Nodejs",
-  "shouldAddHelpers": true,
-  "shouldAddSourcemapSupport": true,
-  "awsLambdaHandler": ""
-}
