--- .vercel/output/functions/src/index.js.func/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "node-express-mongodb-server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-"scripts": {
-  "build": "babel src --out-dir dist",
-  "start": "node dist/index.js"
-},
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@babel/cli": "^7.27.0",
-    "@babel/core": "^7.9.0",
-    "@babel/node": "^7.8.7",
-    "@babel/preset-env": "^7.26.9",
-    "nodemon": "^2.0.3"
-  },
-  "dependencies": {
-    "axios": "^1.7.7",
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.20.3",
-    "chargebee": "^2.45.0",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "fs": "^0.0.1-security",
-    "jsonwebtoken": "^9.0.2",
-    "mongoose": "^5.13.23",
-    "nba": "^4.15.0",
-    "nba-stats-client": "^1.0.2",
-    "stripe": "^17.4.0"
-  }
-}
