--- .vercel/output/functions/src/index.js.func/src/index.js
@@ -1,34 +1,0 @@
-"use strict";
-
-require("dotenv/config");
-var _cors = _interopRequireDefault(require("cors"));
-var _express = _interopRequireDefault(require("express"));
-var _bodyParser = _interopRequireDefault(require("body-parser"));
-var _chargebee = _interopRequireDefault(require("chargebee"));
-var _stripe = _interopRequireDefault(require("stripe"));
-var _models = _interopRequireWildcard(require("./models"));
-var _routes = _interopRequireDefault(require("./routes"));
-function _getRequireWildcardCache(e) { if ("function" != typeof WeakMap) return null; var r = new WeakMap(), t = new WeakMap(); return (_getRequireWildcardCache = function (e) { return e ? t : r; })(e); }
-function _interopRequireWildcard(e, r) { if (!r && e && e.__esModule) return e; if (null === e || "object" != typeof e && "function" != typeof e) return { default: e }; var t = _getRequireWildcardCache(r); if (t && t.has(e)) return t.get(e); var n = { __proto__: null }, a = Object.defineProperty && Object.getOwnPropertyDescriptor; for (var u in e) if ("default" !== u && {}.hasOwnProperty.call(e, u)) { var i = a ? Object.getOwnPropertyDescriptor(e, u) : null; i && (i.get || i.set) ? Object.defineProperty(n, u, i) : n[u] = e[u]; } return n.default = e, t && t.set(e, n), n; }
-function _interopRequireDefault(e) { return e && e.__esModule ? e : { default: e }; }
-const app = (0, _express.default)();
-_chargebee.default.configure({
-  site: process.env.CHARGEBEE_SITE,
-  api_key: process.env.CHARGEBEE_API_KEY
-});
-app.use((0, _cors.default)());
-app.use(_express.default.json());
-app.use(_express.default.urlencoded({
-  extended: true
-}));
-app.use(_bodyParser.default.json());
-app.use(_bodyParser.default.urlencoded({
-  extended: true
-}));
-app.use('/player', _routes.default.player);
-app.use('/subscription', _routes.default.subscription);
-app.use('/user', _routes.default.user);
-(0, _models.connectDb)().then(async () => {
-  app.listen(process.env.PORT, '0.0.0.0', () => console.log(`NBA App listening on port ${process.env.PORT}!`));
-});
-//# sourceMappingURL=index.js.map
