--- .vercel/output/functions/src/index.js.func/src/routes/index.js
@@ -1,16 +1,0 @@
-"use strict";
-
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-exports.default = void 0;
-var _user = _interopRequireDefault(require("./user"));
-var _player = _interopRequireDefault(require("./player"));
-var _subscription = _interopRequireDefault(require("./subscription"));
-function _interopRequireDefault(e) { return e && e.__esModule ? e : { default: e }; }
-var _default = exports.default = {
-  player: _player.default,
-  subscription: _subscription.default,
-  user: _user.default
-};
-//# sourceMappingURL=index.js.map
