+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,7 @@
+{
+  "ExpandedNodes": [
+    ""
+  ],
+  "SelectedNode": "\\C:\\Users\\dally\\Source\\Repos\\NBA_backend",
+  "PreviewInSolutionExplorer": false
+}
