+++ src/models/user.js
@@ -1,4 +1,6 @@
-const mongoose = require("mongoose");
+import mongoose from 'mongoose';
+
+//const mongoose = require("mongoose");
 
 const userSchema = new mongoose.Schema(
   {
@@ -21,4 +23,5 @@ userSchema.methods.isTrialExpired = function () {
 
 const User = mongoose.model("User", userSchema);
 
-module.exports = User;
+//module.exports = User;
+export default User;
